fix; norm ok in execution dir excep excute.c
