--- srcs/execution/init.c
@@ -1,22 +1,0 @@
-#include "minishell.h"
-
-// void	init_conds(t_cond *conds, t_num *num)
-// {
-// 	size_t	i;
-// 	size_t	j;
-
-// 	i = 0;
-// 	while (conds[i].type != TAIL)
-// 	{
-// 		j = 0;
-// 		while (conds[i].cmds[j].type != TAIL)
-// 		{
-// 			num->pipe_in[0] = 0;
-// 			num->pipe_in[1] = -1;
-// 			num->pipe_out[0] = -1;
-// 			num->pipe_out[1] = 1;
-// 			j++;
-// 		}
-// 		i++;
-// 	}
-// }
